requirements.txt
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.3
blosc==1.10.2
cffi==1.14.4
click==7.1.2
cycler==0.10.0
dbimage
decorator==4.4.2
defusedxml==0.6.0
dtoolbioimage==0.1.10
dtoolcore==3.18.0
entrypoints==0.3
imageio==2.9.0
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
notebook==6.2.0
numpy==1.19.5
packaging==20.8
pandas==1.2.1
pandocfilters==1.4.3
parse==1.19.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.14
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
pyzmq==21.0.2
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
mruntools
scikit-image==0.18.1
scipy==1.6.0
Send2Trash==1.5.0
SimpleITK==2.0.2
six==1.15.0
stacktools
terminado==0.9.2
testpath==0.4.4
tifffile==2021.1.14
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
